scripts/requirements.txt
# LiveKit AI Agent Dependencies
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.6.0

# HTTP client for API calls
httpx>=0.27.0

# Additional utilities
python-dotenv>=1.0.0
